refactor: move Member DTOs to member.dto package for better organization
