+++ src/main/java/com/libraryman_api/member/MemberService.java
@@ -12,6 +12,9 @@ import org.springframework.stereotype.Service;
 import com.libraryman_api.exception.InvalidPasswordException;
 import com.libraryman_api.exception.InvalidSortFieldException;
 import com.libraryman_api.exception.ResourceNotFoundException;
+import com.libraryman_api.member.dto.MembersDto;
+import com.libraryman_api.member.dto.UpdateMembersDto;
+import com.libraryman_api.member.dto.UpdatePasswordDto;
 import com.libraryman_api.notification.NotificationService;
 import com.libraryman_api.security.config.PasswordEncoder;
